fix: configure log source for `podman events`
When we switched the log driver from `journald` to `k8s-file` due to
podman being unable to log to the system journal, it broke the events
command since it defaults to searching the journal for log entries.

This commit changes the `events_logger` configuration so it searches for
events in the correct place.